max prj ok

+++ entitySystem/entitySystem/cJSON/cJSON.h
@@ -54,7 +54,7 @@ typedef struct cJSON {
 } cJSON;
 
 typedef struct cJSON_Hooks {
-      void *(*malloc_fn)(size_t sz);
+      void *(*malloc_fn)(int sz);
       void (*free_fn)(void *ptr);
 } cJSON_Hooks;
